requirements.txt
pandas
numpy
Pillow
matplotlib
tensorflow
seaborn
visualkeras
scikit-learn